// types/movie.ts
export type Movie = {
  video: boolean;
  vote_average: number;
  title: string;
  overview: string;
  release_date: string;
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  vote_count: number;
  original_language: string;
  original_title: string;
  poster_path: string;
  id: number;
  popularity: number;
  media_type: string;
};